// src/app/models/updateuser.model.ts
export interface EditUser {
    // UserId: string
    Name:string
    EmailId: string;
    PhoneNumber: string;
    CreatedOn: string;
    ApprovedBy: string;
    ApprovedOn: string;
    Status:string;
    HospitalId: string;
    DepartmentId: string;
    Username: string;
    Password: string;
    Role: string; 
}